Feat: Answer delete 기능에서 접근하는 사용자가 해당 answer을 작성한 사용자가 맞는지 확인하는 코드 추가

+++ src/main/java/com/encore/playground/domain/qna/controller/AnswerController.java
@@ -88,7 +88,10 @@ public class AnswerController {
     @PostMapping("/answer/delete")
     private List<AnswerDto> deleteAnswer(@RequestBody AnswerDeleteDto answerDeleteDto, HttpServletRequest request) {
         MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
-
-        return answerService.deleteAnswer(answerDeleteDto, memberIdDto);
+        if (answerService.isAnswerWriter(answerDeleteDto.getId(), memberIdDto)) {
+            return answerService.deleteAnswer(answerDeleteDto, memberIdDto);
+        } else {
+            return answerService.answerList(answerDeleteDto.getQuestionId());
+        }
     }
 }
